Fixes Dodgy Dimensions and add Create Compat
