+++ forge/src/main/java/whocraft/tardis_refined/neoforge/CommonBus.java
@@ -15,6 +15,7 @@ import whocraft.tardis_refined.ControlGroupCheckers;
 import whocraft.tardis_refined.TardisRefined;
 import whocraft.tardis_refined.command.TardisRefinedCommand;
 import whocraft.tardis_refined.common.crafting.astral_manipulator.ManipulatorRecipes;
+import whocraft.tardis_refined.common.dimension.DimensionHandler;
 import whocraft.tardis_refined.common.dimension.TardisTeleportData;
 import whocraft.tardis_refined.common.hum.TardisHums;
 import whocraft.tardis_refined.common.tardis.TardisDesktops;
@@ -34,6 +35,15 @@ public class CommonBus {
         }
 
     }
+
+    @SubscribeEvent
+    public static void onServerStart(ServerStartedEvent event) {
+        // Load Levels
+        ServerLevel world = event.getServer().getLevel(Level.OVERWORLD);
+        DimensionHandler.loadLevels(world);
+
+    }
+
 
     @SubscribeEvent
     public static void onDatapack(AddReloadListenerEvent addReloadListenerEvent) {
